stt_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
openai-whisper==20240930
numpy<2.0
torch>=2.0.0
python-multipart==0.0.9
websockets==12.0
